[#1554] Refactor Utils to use module extends
Moves utility methods from the main `Onetime::Utils` module into their
own sub-modules (`Sanitation`, `TimeUtils`, `Numbers`, `Strings`,
`Enumerables`). This improves organization and allows for more explicit
dependency management.
