+++ apps/api/v2/models/secret.rb
@@ -38,7 +38,7 @@ module V2
 
     def natural_duration
       # Colloquial representation of the TTL. e.g. "1 day"
-      OT::Utils::TimeUtils.natural_duration lifespan
+      OT::TimeUtils.natural_duration lifespan
     end
     alias :natural_ttl :natural_duration
